ssh-key:create should add key to local ssh-agent.
