+++ src/Command/Api/ApiCommandHelper.php
@@ -548,8 +548,6 @@ class ApiCommandHelper {
       'accounts:drush-aliases',
       // Skip any command that has a duplicative corresponding ACLI command.
       'ide:create',
-      'ide:delete',
-      'ide:list',
       'log:tail',
       'ssh-key:create',
       'ssh-key:create-upload',
