+++ src/Command/Ide/IdeCreateCommand.php
@@ -115,7 +115,7 @@ class IdeCreateCommand extends IdeCommandBase {
     }
 
     $loop = Loop::get();
-    $spinner = LoopHelper::addSpinnerToLoop($loop, 'Waiting for the IDE to be ready. This can take up to 15 minutes...', $this->output);
+    $spinner = LoopHelper::addSpinnerToLoop($loop, 'Waiting for the IDE to be ready. This usually takes 2 - 15 minutes.', $this->output);
 
     $loop->addPeriodicTimer(5, function () use ($loop, $spinner) {
       try {
